# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Car Wash - Company (Empresa) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /company/profile:
    get:
      operationId: getCompanyProfile
      tags: [profile]
      summary: Get the authenticated company profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
    put:
      operationId: updateCompanyProfile
      tags: [profile]
      summary: Update name / phone of the company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyProfileUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyProfile"
...
  /company/availability:
    get:
      operationId: getCompanyAvailability
      tags: [availability]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyAvailability"
    put:
      operationId: updateCompanyAvailability
      tags: [availability]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyAvailabilityUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyAvailability"
  /company/packages:
    get:
      operationId: listCompanyPackages
      tags: [packages]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyPackageOption"
    put:
      operationId: updateCompanyPackages
      tags: [packages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyPackagesUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyPackageOption"
  /company/services:
    get:
      operationId: listCompanyServices
      tags: [services]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyServiceOption"
    put:
      operationId: updateCompanyServices
      tags: [services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyServicesUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyServiceOption"
  /company/earnings:
    get:
      operationId: getCompanyEarnings
      tags: [earnings]
      summary: Earnings summary (wow endpoint)
      parameters:
        - in: query
          name: dateFrom
          schema: { type: string }
        - in: query
          name: dateTo
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyEarnings"
  /company/earnings/services:
    get:
      operationId: listCompanyEarningServices
      tags: [earnings]
      summary: Paginated list of billing line items
      parameters:
        - in: query
          name: dateFrom
          schema: { type: string }
        - in: query
          name: dateTo
          schema: { type: string }
        - in: query
          name: paymentStatus
          schema:
            type: string
            enum: [pendiente, pagado]
        - in: query
          name: paymentType
          schema:
            type: string
            enum: [directo, membresia]
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyBillingList"
components:
  schemas:
...
    CatalogRef:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
...
    CompanyProfile:
      type: object
      required: [id, name, email, phone, active, totalBookings, completedBookings, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        active: { type: boolean }
        rating: { type: ["number", "null"] }
        totalBookings: { type: integer }
        completedBookings: { type: integer }
        createdAt: { type: string }
    CompanyProfileUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
...
    Pagination:
      type: object
      required: [total, page, limit, totalPages]
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
...
    CompanySlot:
      type: object
      required: [time, enabled]
      properties:
        time: { type: string }
        enabled: { type: boolean }
    CompanyAvailability:
      type: object
      required: [slots, blockedDates]
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/CompanySlot"
        blockedDates:
          type: array
          items: { type: string }
    CompanyAvailabilityUpdate:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/CompanySlot"
        blockedDates:
          type: array
          items: { type: string }
    CompanyPackageOption:
      type: object
      required: [package, name, washType, active, color]
      properties:
        package:
          $ref: "#/components/schemas/CatalogRef"
        name: { type: string }
        color: { type: string }
        washType:
          $ref: "#/components/schemas/CatalogRef"
        active: { type: boolean }
    CompanyPackageToggle:
      type: object
      required: [packageId, active]
      properties:
        packageId: { type: string }
        active: { type: boolean }
    CompanyPackagesUpdate:
      type: object
      required: [packages]
      properties:
        packages:
          type: array
          items:
            $ref: "#/components/schemas/CompanyPackageToggle"
    CompanyServiceOption:
      type: object
      required: [service, name, price, active, kind]
      properties:
        service:
          $ref: "#/components/schemas/CatalogRef"
        name: { type: string }
        price: { type: number }
        active: { type: boolean }
        kind:
          type: string
          enum: [wash_type, add_on]
    CompanyServiceToggle:
      type: object
      required: [serviceId, active]
      properties:
        serviceId: { type: string }
        active: { type: boolean }
    CompanyServicesUpdate:
      type: object
      required: [services]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/CompanyServiceToggle"
    CompanyEarnings:
      type: object
      required:
        - dateFrom
        - dateTo
        - totalServices
        - totalAmount
        - pending
        - paid
        - byType
        - dailySeries
      properties:
        dateFrom: { type: string }
        dateTo: { type: string }
        totalServices: { type: integer }
        totalAmount: { type: number }
        pending: { type: number }
        paid: { type: number }
        byType:
          type: object
          required: [directo, membresia]
          properties:
            directo: { type: number }
            membresia: { type: number }
        dailySeries:
          type: array
          items:
            type: object
            required: [date, amount, services]
            properties:
              date: { type: string }
              amount: { type: number }
              services: { type: integer }
    CompanyBilling:
      type: object
      required:
        - id
        - bookingId
        - date
        - time
        - clientName
        - vehicleSize
        - washType
        - addOns
        - amount
        - paymentType
        - paymentStatus
      properties:
        id: { type: string }
        bookingId: { type: string }
        date: { type: string }
        time: { type: string }
        clientName: { type: string }
        vehicleSize:
          $ref: "#/components/schemas/CatalogRef"
        washType:
          $ref: "#/components/schemas/CatalogRef"
        addOns:
          type: array
          items:
            $ref: "#/components/schemas/CatalogRef"
        amount: { type: number }
        paymentType:
          type: string
          enum: [directo, membresia]
        paymentStatus:
          type: string
          enum: [pendiente, pagado]
        paidAt: { type: ["string", "null"] }
    CompanyBillingList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CompanyBilling"
        pagination:
          $ref: "#/components/schemas/Pagination"