# lib/api-zod/src/generated/types/companyPackageToggle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyPackageToggle {
  packageId: string;
  active: boolean;
}