# lib/api-zod/src/generated/types/companyProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Car Wash - Company (Empresa) API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyProfile {
  id: string;
  name: string;
  email: string;
  phone: string;
  active: boolean;
  /** @nullable */
  rating?: number | null;
  totalBookings: number;
  completedBookings: number;
  createdAt: string;
}